refactor: api dto 수정
