+++ apps/commerce-api/src/main/java/com/loopers/application/member/MemberFacade.java
@@ -8,7 +8,7 @@ import com.loopers.application.provided.MemberFinder;
 import com.loopers.application.provided.MemberRegister;
 import com.loopers.domain.member.Member;
 import com.loopers.domain.member.MemberId;
-import com.loopers.interfaces.api.member.dto.MemberRegisterRequest;
+import com.loopers.interfaces.api.member.dto.MemberV1Dto.Request.MemberRegisterRequest;
 
 import lombok.RequiredArgsConstructor;
