+++ apps/commerce-api/src/main/java/com/loopers/application/member/MemberModifyService.java
@@ -12,7 +12,7 @@ import com.loopers.domain.member.DuplicateMemberIdException;
 import com.loopers.domain.member.Member;
 import com.loopers.domain.member.CreateMemberSpec;
 import com.loopers.domain.member.MemberId;
-import com.loopers.interfaces.api.member.dto.MemberRegisterRequest;
+import com.loopers.interfaces.api.member.dto.MemberV1Dto.Request.MemberRegisterRequest;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
