+++ apps/commerce-api/src/main/java/com/loopers/application/provided/MemberRegister.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 import com.loopers.domain.member.Member;
 import com.loopers.domain.member.MemberId;
-import com.loopers.interfaces.api.member.dto.MemberRegisterRequest;
+import com.loopers.interfaces.api.member.dto.MemberV1Dto.Request.MemberRegisterRequest;
 
 import jakarta.validation.Valid;
