+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/member/MemberV1ApiController.java
@@ -10,9 +10,9 @@ import com.loopers.application.member.MemberFacade;
 import com.loopers.domain.member.Member;
 import com.loopers.domain.member.MemberId;
 import com.loopers.interfaces.api.ApiResponse;
-import com.loopers.interfaces.api.member.dto.MemberRegisterRequest;
+import com.loopers.interfaces.api.member.dto.MemberV1Dto.Request.MemberRegisterRequest;
 import com.loopers.interfaces.api.member.dto.MemberV1Dto;
-import com.loopers.interfaces.api.member.dto.MemberV1Dto.MemberInfoResponse;
+import com.loopers.interfaces.api.member.dto.MemberV1Dto.Response.MemberInfoResponse;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -25,16 +25,16 @@ public class MemberV1ApiController implements MemberV1ApiSpec {
 
     @PostMapping
     @Override
-    public ApiResponse<MemberV1Dto.MemberRegisterResponse> register(@RequestBody @Valid MemberRegisterRequest registerRequest) {
+    public ApiResponse<MemberV1Dto.Response.MemberRegisterResponse> register(@RequestBody @Valid MemberRegisterRequest registerRequest) {
         Member member = memberFacade.register(registerRequest);
 
-        MemberV1Dto.MemberRegisterResponse memberRegisterResponse = MemberV1Dto.MemberRegisterResponse.of(member);
+        MemberV1Dto.Response.MemberRegisterResponse memberRegisterResponse = MemberV1Dto.Response.MemberRegisterResponse.of(member);
         return ApiResponse.success(memberRegisterResponse);
     }
 
     @GetMapping("/me")
     @Override
-    public ApiResponse<MemberV1Dto.MemberInfoResponse> find(MemberId memberId) {
+    public ApiResponse<MemberV1Dto.Response.MemberInfoResponse> find(MemberId memberId) {
         return ApiResponse.success(MemberInfoResponse.of(memberFacade.find(memberId)));
     }
 }
