+++ apps/commerce-api/src/test/java/com/loopers/interfaces/api/member/MemberV1ApiE2ETest.java
@@ -28,7 +28,7 @@ import com.loopers.domain.member.Member;
 import com.loopers.domain.member.MemberFixture;
 import com.loopers.infrastructure.member.MemberJpaRepository;
 import com.loopers.interfaces.api.ApiResponse;
-import com.loopers.interfaces.api.member.dto.MemberRegisterRequest;
+import com.loopers.interfaces.api.member.dto.MemberV1Dto.Request.MemberRegisterRequest;
 import com.loopers.interfaces.api.member.dto.MemberV1Dto;
 import com.loopers.utils.DatabaseCleanUp;
 
@@ -70,9 +70,9 @@ class MemberV1ApiE2ETest {
             HttpHeaders headers = new HttpHeaders();
             headers.setContentType(MediaType.APPLICATION_JSON);
 
-            ParameterizedTypeReference<ApiResponse<MemberV1Dto.MemberRegisterResponse>> responseType =
+            ParameterizedTypeReference<ApiResponse<MemberV1Dto.Response.MemberRegisterResponse>> responseType =
                     new ParameterizedTypeReference<>() {};
-            ResponseEntity<ApiResponse<MemberV1Dto.MemberRegisterResponse>> response =
+            ResponseEntity<ApiResponse<MemberV1Dto.Response.MemberRegisterResponse>> response =
                     testRestTemplate.exchange(ENDPOINT_POST,
                                               HttpMethod.POST,
                                               new HttpEntity<>(memberRegisterJson, headers),
@@ -97,9 +97,9 @@ class MemberV1ApiE2ETest {
             HttpHeaders headers = new HttpHeaders();
             headers.setContentType(MediaType.APPLICATION_JSON);
 
-            ParameterizedTypeReference<ApiResponse<MemberV1Dto.MemberRegisterResponse>> responseType =
+            ParameterizedTypeReference<ApiResponse<MemberV1Dto.Response.MemberRegisterResponse>> responseType =
                     new ParameterizedTypeReference<>() {};
-            ResponseEntity<ApiResponse<MemberV1Dto.MemberRegisterResponse>> response =
+            ResponseEntity<ApiResponse<MemberV1Dto.Response.MemberRegisterResponse>> response =
                     testRestTemplate.exchange(ENDPOINT_POST,
                                               HttpMethod.POST,
                                               new HttpEntity<>(memberRegisterJson, headers),
@@ -121,9 +121,9 @@ class MemberV1ApiE2ETest {
         void get_memberInfo() {
             Member member = memberJpaRepository.saveAndFlush(MemberFixture.createMember());
 
-            ParameterizedTypeReference<ApiResponse<MemberV1Dto.MemberInfoResponse>> responseType =
+            ParameterizedTypeReference<ApiResponse<MemberV1Dto.Response.MemberInfoResponse>> responseType =
                     new ParameterizedTypeReference<>() {};
-            ResponseEntity<ApiResponse<MemberV1Dto.MemberInfoResponse>> response =
+            ResponseEntity<ApiResponse<MemberV1Dto.Response.MemberInfoResponse>> response =
                     testRestTemplate.exchange(RequestEntity.get(ENDPOINT_GET)
                                                            .header("X-USER-ID", member.getMemberId().memberId())
                                                            .build(),
